chore(chstorage): improve metrics querying tracing
