+++ internal/chstorage/querier_metrics.go
@@ -125,401 +125,6 @@ func DecodeUnicodeLabel(v string) string {
 	return sb.String()
 }
 
-// LabelValues returns all potential values for a label name.
-// It is not safe to use the strings beyond the lifetime of the querier.
-// If matchers are specified the returned result set is reduced
-// to label values of metrics matching the matchers.
-func (p *promQuerier) LabelValues(ctx context.Context, labelName string, hints *storage.LabelHints, matchers ...*labels.Matcher) (result []string, _ annotations.Annotations, rerr error) {
-	if hints == nil {
-		hints = &storage.LabelHints{}
-	}
-	labelName = DecodeUnicodeLabel(labelName)
-	ctx, span := p.tracer.Start(ctx, "chstorage.metrics.LabelValues",
-		trace.WithAttributes(
-			attribute.String("chstorage.label", labelName),
-			attribute.Int("chstorage.hints.limits", hints.Limit),
-			xattribute.StringerSlice("chstorage.matchers", matchers),
-			xattribute.UnixNano("chstorage.mint", p.mint),
-			xattribute.UnixNano("chstorage.maxt", p.maxt),
-		),
-	)
-	defer func() {
-		if rerr != nil {
-			span.RecordError(rerr)
-		} else {
-			span.AddEvent("values_fetched", trace.WithAttributes(
-				attribute.Int("chstorage.total_values", len(result)),
-			))
-		}
-		span.End()
-	}()
-
-	matchesOtherLabels := slices.ContainsFunc(matchers, func(m *labels.Matcher) bool {
-		return labelName != m.Name
-	})
-	if matchesOtherLabels {
-		r, err := p.getMatchingLabelValues(ctx, labelName, matchers)
-		if err != nil {
-			return nil, nil, err
-		}
-		return r, nil, nil
-	}
-
-	r, err := p.getLabelValues(ctx, labelName, matchers...)
-	if err != nil {
-		return nil, nil, err
-	}
-	return r, nil, nil
-}
-
-func (p *promQuerier) getLabelValues(ctx context.Context, labelName string, matchers ...*labels.Matcher) (result []string, rerr error) {
-	table := p.tables.Labels
-
-	ctx, span := p.tracer.Start(ctx, "chstorage.metrics.getLabelValues",
-		trace.WithAttributes(
-			attribute.String("chstorage.label", labelName),
-			xattribute.StringerSlice("chstorage.matchers", matchers),
-
-			attribute.String("chstorage.table", table),
-		),
-	)
-	defer func() {
-		if rerr != nil {
-			span.RecordError(rerr)
-		}
-		span.End()
-	}()
-
-	var (
-		value       proto.ColStr
-		valueColumn = "value"
-	)
-
-	query := chsql.Select(table, chsql.Column(valueColumn, &value)).
-		Distinct(true).
-		Where(chsql.ColumnEq("name", labelName))
-	for _, m := range matchers {
-		if m.Name != labelName {
-			return nil, errors.Errorf("unexpected label matcher %s (label must be %q)", m, labelName)
-		}
-
-		expr, err := promQLLabelMatcher(
-			[]chsql.Expr{chsql.Ident(valueColumn)},
-			m.Type,
-			m.Value,
-		)
-		if err != nil {
-			return nil, err
-		}
-		query.Where(expr)
-	}
-	query.Order(chsql.Ident(valueColumn), chsql.Asc)
-
-	if err := p.do(ctx, selectQuery{
-		Query: query,
-		OnResult: func(ctx context.Context, block proto.Block) error {
-			for i := 0; i < value.Rows(); i++ {
-				result = append(result, value.Row(i))
-			}
-			return nil
-		},
-
-		Type:   "getLabelValues",
-		Signal: "metrics",
-		Table:  table,
-	}); err != nil {
-		return nil, err
-	}
-
-	return result, nil
-}
-
-func (p *promQuerier) getMatchingLabelValues(ctx context.Context, labelName string, matchers []*labels.Matcher) (_ []string, rerr error) {
-	table := p.tables.Timeseries
-
-	ctx, span := p.tracer.Start(ctx, "chstorage.metrics.getMatchingLabelValues",
-		trace.WithAttributes(
-			attribute.String("chstorage.label", labelName),
-			xattribute.StringerSlice("chstorage.matchers", matchers),
-		),
-	)
-	defer func() {
-		if rerr != nil {
-			span.RecordError(rerr)
-		}
-		span.End()
-	}()
-
-	mlabels := make([]string, 0, len(matchers))
-	for _, m := range matchers {
-		mlabels = append(mlabels, m.Name)
-	}
-	mlabels = append(mlabels, labelName)
-	mapping, err := p.getLabelMapping(ctx, mlabels)
-	if err != nil {
-		return nil, errors.Wrap(err, "get label mapping")
-	}
-
-	var (
-		value      proto.ColumnOf[string]
-		columnExpr chsql.Expr
-	)
-	if labelName == labels.MetricName {
-		columnExpr = chsql.Ident("name")
-		value = proto.NewLowCardinality(&proto.ColStr{})
-	} else {
-		columnExpr = firstAttrSelector(labelName)
-		value = &proto.ColStr{}
-	}
-
-	query := chsql.Select(table, chsql.ResultColumn{
-		Name: "value",
-		Expr: columnExpr,
-		Data: value,
-	}).
-		Distinct(true)
-
-	for _, m := range matchers {
-		selectors := []chsql.Expr{
-			chsql.Ident("name"),
-		}
-		if name := m.Name; name != labels.MetricName {
-			selectors = mapping.Selectors(name)
-		}
-
-		expr, err := promQLLabelMatcher(selectors, m.Type, m.Value)
-		if err != nil {
-			return nil, err
-		}
-		query.Where(expr)
-	}
-	query.GroupBy(
-		chsql.Ident("name"),
-		chsql.Ident("attribute"),
-		chsql.Ident("scope"),
-		chsql.Ident("resource"),
-	)
-	if !p.mint.IsZero() {
-		query.Having(chsql.Gte(
-			chsql.ToUnixTimestamp64Nano(chsql.Ident("first_seen")),
-			chsql.UnixNano(p.mint),
-		))
-	}
-	if !p.maxt.IsZero() {
-		query.Having(chsql.Lte(
-			chsql.ToUnixTimestamp64Nano(chsql.Ident("last_seen")),
-			chsql.UnixNano(p.maxt),
-		))
-	}
-	query.Limit(p.labelLimit)
-
-	var result []string
-	if err := p.do(ctx, selectQuery{
-		Query: query,
-		OnResult: func(ctx context.Context, block proto.Block) error {
-			for i := 0; i < value.Rows(); i++ {
-				if v := value.Row(i); v != "" {
-					result = append(result, v)
-				}
-			}
-			return nil
-		},
-
-		Type:   "getMatchingLabelValues",
-		Signal: "metrics",
-		Table:  table,
-	}); err != nil {
-		return nil, err
-	}
-	span.AddEvent("values_fetched", trace.WithAttributes(
-		attribute.String("chstorage.table", table),
-		attribute.Int("chstorage.total_values", len(result)),
-	))
-
-	slices.Sort(result)
-	// Remove duplicates.
-	result = slices.Clip(result)
-
-	return result, nil
-}
-
-// LabelNames returns all the unique label names present in the block in sorted order.
-// If matchers are specified the returned result set is reduced
-// to label names of metrics matching the matchers.
-func (p *promQuerier) LabelNames(ctx context.Context, hints *storage.LabelHints, matchers ...*labels.Matcher) (result []string, _ annotations.Annotations, rerr error) {
-	if hints == nil {
-		hints = &storage.LabelHints{}
-	}
-	ctx, span := p.tracer.Start(ctx, "chstorage.metrics.LabelNames",
-		trace.WithAttributes(
-			xattribute.StringerSlice("chstorage.matchers", matchers),
-			attribute.Int("chstorage.hints.limit", hints.Limit),
-			xattribute.UnixNano("chstorage.mint", p.mint),
-			xattribute.UnixNano("chstorage.maxt", p.maxt),
-		),
-	)
-	defer func() {
-		if rerr != nil {
-			span.RecordError(rerr)
-		} else {
-			span.AddEvent("names_fetched", trace.WithAttributes(
-				attribute.Int("chstorage.total_names", len(result)),
-			))
-		}
-		span.End()
-	}()
-
-	var err error
-	if len(matchers) > 0 {
-		result, err = p.getMatchingLabelNames(ctx, matchers)
-		if err != nil {
-			return nil, nil, err
-		}
-		return result, nil, nil
-	}
-
-	result, err = p.getLabelNames(ctx)
-	if err != nil {
-		return nil, nil, err
-	}
-	return result, nil, nil
-}
-
-func (p *promQuerier) getLabelNames(ctx context.Context) (result []string, rerr error) {
-	table := p.tables.Labels
-
-	ctx, span := p.tracer.Start(ctx, "chstorage.metrics.getLabelNames",
-		trace.WithAttributes(
-			attribute.String("chstorage.table", table),
-		),
-	)
-	defer func() {
-		if rerr != nil {
-			span.RecordError(rerr)
-		}
-		span.End()
-	}()
-
-	var (
-		value = new(proto.ColStr).LowCardinality()
-		query = chsql.Select(table, chsql.Column("name", value)).
-			Distinct(true).
-			Order(chsql.Ident("name"), chsql.Asc)
-	)
-	if err := p.do(ctx, selectQuery{
-		Query: query,
-		OnResult: func(ctx context.Context, block proto.Block) error {
-			for i := 0; i < value.Rows(); i++ {
-				result = append(result, value.Row(i))
-			}
-			return nil
-		},
-
-		Type:   "LabelNames",
-		Signal: "metrics",
-		Table:  table,
-	}); err != nil {
-		return nil, err
-	}
-
-	return result, nil
-}
-
-func (p *promQuerier) getMatchingLabelNames(ctx context.Context, matchers []*labels.Matcher) (_ []string, rerr error) {
-	table := p.tables.Timeseries
-
-	ctx, span := p.tracer.Start(ctx, "chstorage.metrics.getMatchingLabelNames",
-		trace.WithAttributes(
-			xattribute.StringerSlice("chstorage.matchers", matchers),
-		),
-	)
-	defer func() {
-		if rerr != nil {
-			span.RecordError(rerr)
-		}
-		span.End()
-	}()
-
-	mlabels := make([]string, 0, len(matchers))
-	for _, m := range matchers {
-		mlabels = append(mlabels, m.Name)
-	}
-	mapping, err := p.getLabelMapping(ctx, mlabels)
-	if err != nil {
-		return nil, errors.Wrap(err, "get label mapping")
-	}
-
-	var (
-		value proto.ColStr
-
-		query = chsql.Select(table, chsql.ResultColumn{
-			Name: "values",
-			Expr: chsql.ArrayJoin(
-				chsql.ArrayConcat(
-					attrKeys(colAttrs),
-					attrKeys(colScope),
-					attrKeys(colResource),
-				),
-			),
-			Data: &value,
-		}).
-			Distinct(true)
-	)
-	for _, m := range matchers {
-		selectors := []chsql.Expr{
-			chsql.Ident("name"),
-		}
-		if name := m.Name; name != labels.MetricName {
-			selectors = mapping.Selectors(name)
-		}
-
-		expr, err := promQLLabelMatcher(selectors, m.Type, m.Value)
-		if err != nil {
-			return nil, err
-		}
-		query.Where(expr)
-	}
-	query.GroupBy(
-		chsql.Ident("name"),
-		chsql.Ident("attribute"),
-		chsql.Ident("scope"),
-		chsql.Ident("resource"),
-	)
-	query.Limit(p.labelLimit)
-
-	var result []string
-	if err := p.do(ctx, selectQuery{
-		Query: query,
-		OnResult: func(ctx context.Context, block proto.Block) error {
-			for i := 0; i < value.Rows(); i++ {
-				result = append(result, value.Row(i))
-			}
-			return nil
-		},
-
-		Type:   "getMatchingLabelNames",
-		Signal: "metrics",
-		Table:  table,
-	}); err != nil {
-		return nil, err
-	}
-	span.AddEvent("labels_fetched", trace.WithAttributes(
-		attribute.String("chstorage.table", table),
-		attribute.Int("chstorage.total_values", len(result)),
-	))
-
-	if len(result) > 0 {
-		// Add `__name__` only if there is any matching series.
-		result = append(result, labels.MetricName)
-	}
-
-	slices.Sort(result)
-	// Remove duplicates.
-	result = slices.Clip(result)
-
-	return result, nil
-}
-
 func promQLLabelMatcher(valueSel []chsql.Expr, typ labels.MatchType, value string) (e chsql.Expr, rerr error) {
 	defer func() {
 		if rerr == nil {
@@ -768,6 +373,9 @@ func (q *Querier) queryMetricsTimeseries(
 	}); err != nil {
 		return nil, err
 	}
+	span.AddEvent("timeseries_fetched", trace.WithAttributes(
+		attribute.Int("chstorage.total_series", len(set)),
+	))
 
 	return set, nil
 }
@@ -894,7 +502,21 @@ func (p *promQuerier) extractHints(
 	return hints, start, end, mlabels
 }
 
-func (p *promQuerier) queryPoints(ctx context.Context, table string, start, end time.Time, timeseries map[[16]byte]labels.Labels) ([]storage.Series, error) {
+func (p *promQuerier) queryPoints(ctx context.Context, table string, start, end time.Time, timeseries map[[16]byte]labels.Labels) (_ []storage.Series, rerr error) {
+	ctx, span := p.tracer.Start(ctx, "chstorage.metrics.queryPoints",
+		trace.WithAttributes(
+			xattribute.UnixNano("chstorage.range.start", start),
+			xattribute.UnixNano("chstorage.range.end", end),
+			attribute.String("chstorage.table", table),
+		),
+	)
+	defer func() {
+		if rerr != nil {
+			span.RecordError(rerr)
+		}
+		span.End()
+	}()
+
 	var (
 		c     = newPointColumns()
 		query = chsql.Select(table, c.ChsqlResult()...).
@@ -913,7 +535,10 @@ func (p *promQuerier) queryPoints(ctx context.Context, table string, start, end 
 		inputData.Append(hash)
 	}
 
-	set := map[[16]byte]*series[pointData]{}
+	var (
+		set         = map[[16]byte]*series[pointData]{}
+		totalPoints int
+	)
 	if err := p.do(ctx, selectQuery{
 		Query:         query,
 		ExternalTable: "timeseries_hashes",
@@ -942,6 +567,8 @@ func (p *promQuerier) queryPoints(ctx context.Context, table string, start, end 
 
 				s.data.values = append(s.data.values, value)
 				s.ts = append(s.ts, timestamp.UnixMilli())
+
+				totalPoints++
 			}
 			return nil
 		},
@@ -952,6 +579,10 @@ func (p *promQuerier) queryPoints(ctx context.Context, table string, start, end 
 	}); err != nil {
 		return nil, err
 	}
+	span.AddEvent("points_fetched", trace.WithAttributes(
+		attribute.Int("chstorage.total_series", len(set)),
+		attribute.Int("chstorage.total_points", totalPoints),
+	))
 
 	result := make([]storage.Series, 0, len(set))
 	for _, s := range set {
@@ -960,7 +591,21 @@ func (p *promQuerier) queryPoints(ctx context.Context, table string, start, end 
 	return result, nil
 }
 
-func (p *promQuerier) queryExpHistograms(ctx context.Context, table string, start, end time.Time, timeseries map[[16]byte]labels.Labels) ([]storage.Series, error) {
+func (p *promQuerier) queryExpHistograms(ctx context.Context, table string, start, end time.Time, timeseries map[[16]byte]labels.Labels) (_ []storage.Series, rerr error) {
+	ctx, span := p.tracer.Start(ctx, "chstorage.metrics.queryExpHistograms",
+		trace.WithAttributes(
+			xattribute.UnixNano("chstorage.range.start", start),
+			xattribute.UnixNano("chstorage.range.end", end),
+			attribute.String("chstorage.table", table),
+		),
+	)
+	defer func() {
+		if rerr != nil {
+			span.RecordError(rerr)
+		}
+		span.End()
+	}()
+
 	var (
 		c     = newExpHistogramColumns()
 		query = chsql.Select(table, c.ChsqlResult()...).Where(
@@ -977,7 +622,10 @@ func (p *promQuerier) queryExpHistograms(ctx context.Context, table string, star
 		inputData.Append(hash)
 	}
 
-	set := map[[16]byte]*series[expHistData]{}
+	var (
+		set         = map[[16]byte]*series[expHistData]{}
+		totalPoints int
+	)
 	if err := p.do(ctx, selectQuery{
 		Query:         query,
 		ExternalTable: "timeseries_hashes",
@@ -1024,6 +672,8 @@ func (p *promQuerier) queryExpHistograms(ctx context.Context, table string, star
 				s.data.negativeOffset = append(s.data.negativeOffset, negativeOffset)
 				s.data.negativeBucketCounts = append(s.data.negativeBucketCounts, negativeBucketCounts)
 				s.ts = append(s.ts, timestamp.UnixMilli())
+
+				totalPoints++
 			}
 			return nil
 		},
@@ -1034,6 +684,10 @@ func (p *promQuerier) queryExpHistograms(ctx context.Context, table string, star
 	}); err != nil {
 		return nil, err
 	}
+	span.AddEvent("histograms_fetched", trace.WithAttributes(
+		attribute.Int("chstorage.total_series", len(set)),
+		attribute.Int("chstorage.total_points", totalPoints),
+	))
 
 	result := make([]storage.Series, 0, len(set))
 	for _, s := range set {
